% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_ref.R
\docType{data}
\name{cdc_ref}
\alias{cdc_ref}
\title{CDC reference}
\format{A dataframe.
\describe{
  \item{}{Still need to document all variables - there are lots.}
}}
\usage{
cdc_ref
}
\description{
Reference data for CDC data.
}
\keyword{datasets}
