% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_convert.R
\name{convert_kg_lbs}
\alias{convert_kg_lbs}
\title{Kilogram to pound conversion}
\usage{
convert_kg_lbs(kg)
}
\arguments{
\item{kg}{weight in kilograms}
}
\value{
Converted value in pounds
}
\description{
Converts between kilograms into pounds
}
\examples{
convert_kg_lbs(7.123)
#15.70352
convert_kg_lbs(c(7.836,7.432,6.921))
#17.27542 16.38475 15.25819
}
\references{
NIST Guide to the SI, Appendix B.9: Factors for units listed by kind of quantity or field of science https://www.nist.gov/pml/nist-guide-si-appendix-b9-factors-units-listed-kind-quantity-or-field-science
}
