% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_convert.R
\name{convert_lbs_kg}
\alias{convert_lbs_kg}
\title{Pound to kilogram conversion}
\usage{
convert_lbs_kg(lbs)
}
\value{
Converted value in kilograms
}
\description{
Converts between pounds into  kilograms
}
\examples{
convert_lbs_kg(7.123)
#3.230939
convert_lbs_kg(c(7.836,7.432,6.921))
#3.554350 3.371099 3.139313
}
\references{
NIST Guide to the SI, Appendix B.9: Factors for units listed by kind of quantity or field of science https://www.nist.gov/pml/nist-guide-si-appendix-b9-factors-units-listed-kind-quantity-or-field-science
}
