% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_growth.R
\name{create_cdc_growth}
\alias{create_cdc_growth}
\title{create_cdc_growth}
\usage{
create_cdc_growth(df)
}
\arguments{
\item{df}{Input data.frame.}
}
\value{
the original data frame with percentiles, z-scores, and related values
added as additional columns
}
\description{
Calculate important maternal child health measures from a dataset
}
\examples{
## NHANES example data is included with package
create_cdc_growth(nhanes_data)
}
\references{
Cole TJ, Bellizzi MC, Flegal KM, Dietz WH. Establishing a
standard definition for child overweight and obesity worldwide: international
survey. BMJ: British Medical Journal. 2000;320(7244):1240.

https://www.cdc.gov/nccdphp/dnpao/growthcharts/resources/sas.htm
}
