% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_mchtoolbox.R
\name{launch_mchtoolbox}
\alias{launch_mchtoolbox}
\title{Start the Shiny Interface for mchtoolbox}
\usage{
launch_mchtoolbox()
}
\value{
Nothing
}
\description{
startmavis starts loads the web browser an interactive user interface built using R shiny.
}
\details{
Launch mchtoolbox

The purpose of building the interactive user interface is to provide an easy for people who are learning how to do their first meta-analysis and/or are not comfortable with the R command line system. Includes example data for testing out a meta-analysis.
}
\examples{
\dontrun{
library(shiny)
lauch_mchtoolbox()
}
}
\keyword{mchtoolbox}
