% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated example dataset}
\format{A dataframe.
\describe{
  \item{cid}{Patient identifier.}
  \item{sex}{1 = male, 2 = female}
  \item{agemons}{Age in months.}
  \item{height}{Height in cm.}
  \item{weight}{Weight in kg.}
}}
\usage{
sim_data
}
\description{
Simulated data.
}
\keyword{datasets}
