% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdc_growth.R
\name{create_cdc_growth}
\alias{create_cdc_growth}
\title{Title}
\usage{
create_cdc_growth(df)
}
\arguments{
\item{data}{Input data.frame.}
}
\description{
Title
}
\references{
Cole TJ, Bellizzi MC, Flegal KM, Dietz WH. Establishing a
standard definition for child overweight and obesity worldwide: international
survey. BMJ: British Medical Journal. 2000;320(7244):1240.

https://www.cdc.gov/nccdphp/dnpao/growthcharts/resources/sas.htm
}
