% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_mchtoolbox.R
\name{launch_mchtoolbox}
\alias{launch_mchtoolbox}
\title{Start the Shiny Interface for mchtoolbox}
\usage{
launch_mchtoolbox()
}
\value{
Nothing
}
\description{
lauch_mchtoolbox loads an interactive user interface built using R shiny.
}
\details{
Launch mchtoolbox

The purpose of building the interactive user interface is to provide an easy method for people to use instead of the command line.
}
\examples{
\dontrun{
library(shiny)
lauch_mchtoolbox()
}
}
\keyword{mchtoolbox}
