% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_data.R
\docType{data}
\name{nhanes_data}
\alias{nhanes_data}
\title{National Health and Nutrition Examination Survey (NHANES) example dataset}
\format{A dataframe.}
\usage{
nhanes_data
}
\description{
Select data from the US National Health and Nutrition Examination Survey (NHANES). Taken from the NHANES package on CRAN, should only be used for educational purposes  https://cran.r-project.org/web/packages/NHANES https://www.cdc.gov/nchs/nhanes/index.htm
\describe{
  \item{cid}{Patient identifier.}
  \item{sex}{1 = male, 2 = female}
  \item{agemons}{Age in months.}
  \item{height}{Height in cm.}
  \item{weight}{Weight in kg.}
  \item{headcir}{Head circumference in cm.}
}
}
\details{
NHANES data.
}
\keyword{datasets}
