% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanes_data.R
\docType{data}
\name{nhanes_data}
\alias{nhanes_data}
\title{National Health and Nutrition Examination Survey (NHANES) example dataset}
\format{A dataframe.
\describe{
  \item{cid}{Patient identifier.}
  \item{sex}{1 = male, 2 = female}
  \item{agemons}{Age in months.}
  \item{height}{Height in cm.}
  \item{weight}{Weight in kg.}
  \item{headcir}{Head circumference in cm.}
}}
\usage{
nhanes_data
}
\description{
NHANES data.
}
\keyword{datasets}
